cdrom: make the LBA-to-MSF conversions defined for image-supplied LBAs
lba + 150 appears in five places across the CD backends and the shared
utility header. lba comes from the disc image's own tables - a cue's
INDEX values, a CCD entry's PLBA, a CHD or PBP TOC - so a malformed image
reaches these with values at the extremes of int32_t and the addition
overflows a signed int, which is undefined.

    CDUtility.c    subpw_synth_udapp_lba, subpw_synth_leadout_lba
    CDUtility.h    LBA_to_ABA
    CDAccess_Image.c, CDAccess_PBP.c, CDAccess_CHD.c   MakeSubPQ

Each is wrapped through uint32_t and converted back, which is exactly the
result -fwrapv already produces and what the code has always relied on.
The encode_mode*_sector call sites are left alone: those take a uint32_t
parameter, so the conversion there is already defined.

The signed division and remainder are deliberately kept. This is the part
worth being careful about, and a first attempt got it wrong: computing
the whole expression in unsigned looks equivalent and is not. For a
pregap LBA the original produces a negative remainder which then converts
to the uint32_t result, and unsigned modulo gives a different value
entirely - 33.7M differences over 40M samples. Only the addition needed
to change.

Verified against the original semantics rather than against an
idealisation of them: the old expressions were compiled into their own
translation unit with -fwrapv, which is how the core builds, and compared
with the rewrite over 40M random pairs biased toward negatives and the
int32_t extremes. Zero differences.

Found while fuzzing the CCD path, which reached subpw_synth_leadout_lba
first; the rest came from grepping the pattern once the shape was known.
The CCD corpus now runs 2500 iterations clean on the seed that found it
and 600 each on three more, and the cue corpus from the previous commit
still runs 2000 clean. Parsing of both valid images is unchanged - same
TOC, same track LBAs and control bytes, same leadout, same sector read -
compared against a build of the unmodified tree.

Full core build clean; warning counts unchanged against baseline for all
five files.

One harness note, since it cost time: the fuzzer initially reported
nothing on CCD because CDAccess_CCD derives its .img and .sub siblings
from the sheet's own filename, so mutated sheets written under a new name
were rejected before parsing. The 0-findings result that produced was
measuring nothing.

+++ mednafen/cdrom/CDUtility.h
@@ -122,7 +122,10 @@ static INLINE int32_t ABA_to_LBA(uint32_t aba)
 
 static INLINE uint32_t LBA_to_ABA(int32_t lba)
 {
-   return(lba + 150);
+   /* Wrapped through uint32_t: the result is a uint32_t either way, and
+    * lba arrives from image-supplied tables where it can sit at the
+    * extremes of int32_t, making the bare addition undefined. */
+   return((uint32_t)lba + 150u);
 }
 
 static INLINE int32_t AMSF_to_LBA(uint8_t m_a, uint8_t s_a, uint8_t f_a)
